Move all Product Card tests into ProductCard.test.tsx
